move JHipsterModuleGradleProfilePlugins class to javabuildprofile package
