+++ src/main/java/tech/jhipster/lite/module/domain/javabuildprofile/JHipsterModuleJavaBuildProfile.java
@@ -3,9 +3,8 @@ package tech.jhipster.lite.module.domain.javabuildprofile;
 import java.util.Optional;
 import tech.jhipster.lite.module.domain.buildproperties.JHipsterModuleBuildProperties;
 import tech.jhipster.lite.module.domain.buildproperties.JHipsterModuleBuildProperties.JHipsterModuleBuildPropertiesBuilder;
-import tech.jhipster.lite.module.domain.gradleplugin.JHipsterModuleGradleProfilePlugins;
-import tech.jhipster.lite.module.domain.gradleplugin.JHipsterModuleGradleProfilePlugins.JHipsterModuleGradleProfilePluginBuilder;
 import tech.jhipster.lite.module.domain.javabuildprofile.BuildProfileActivation.BuildProfileActivationBuilder;
+import tech.jhipster.lite.module.domain.javabuildprofile.JHipsterModuleGradleProfilePlugins.JHipsterModuleGradleProfilePluginBuilder;
 import tech.jhipster.lite.module.domain.javabuildprofile.JHipsterModuleJavaBuildProfiles.JHipsterModuleJavaBuildProfilesBuilder;
 import tech.jhipster.lite.module.domain.javadependency.JHipsterModuleJavaDependencies;
 import tech.jhipster.lite.module.domain.javadependency.JHipsterModuleJavaDependencies.JHipsterModuleJavaDependenciesBuilder;
